Add completion state
